Redo ack pulse.
Turns out the ack pulse wasn't actually being turned off by the timer (the timer did
nothing), just by the next DCC packet. That worked enough, the timings
worked out (probably on purpose), but it's clearly wrong and not
really reliable. This led to a giant rabbit hole that ended up with the
realization that the Neopixels library uses timing functions that use
timers, and since all (both of them) are used here, that won't work. So as
a workaround (temporary I hope but you know how it is), this is stubbed
out in a way that fools the library. Weirdness abounds, but timer1 is
now used correctly.